refactor(shared): move Sidebar component to shared for reuse
relocated Sidebar component to shared directory to enable reuse
across issue creation and detail views
