+++ frontend/src/shared/components/sidebar/AssigneeSection.tsx
