+++ frontend/src/shared/components/sidebar/MilestoneSection.tsx
