+++ frontend/src/shared/components/sidebar/SidebarSection.tsx
