+++ frontend/src/shared/components/sidebar/index.tsx
